refactor: change design cart page
