+++ src/components/Cart/CartEmpty.tsx
@@ -1,13 +1,12 @@
 import { Button } from 'primereact/button';
 import { useNavigate } from 'react-router-dom';
 import { PAGES } from '../../constants/pages';
-import styles from './CartForm.module.scss';
-// import { count } from '../constants/registratForm';
+import styles from './CartEmpty.module.scss';
 
-export default function CartEmpty(): JSX.Element {
+export function CartEmpty(): JSX.Element {
   const redirect = useNavigate();
   return (
-    <div className={styles.cart_emtpy}>
+    <div className={styles.cart_empty}>
       <p style={{ margin: '0.3rem' }}>Cart Empty</p>
       <Button
         className="mt-3 mb-1 border-round-lg"
