+++ src/components/Links.module.scss
@@ -3,3 +3,11 @@
 .fontSize {
   font-size: 2em !important;
 }
+
+.none {
+  display: none !important;;
+}
+
+.block {
+  display: inline-block;
+}
