+++ src/pages/Cart/Cart.module.scss
@@ -1,7 +1,7 @@
 @import '../../styles/global';
 
 .page {
-  display: flex;
+  display: block;
   min-height: calc(100vh - 150px);
   padding: 20px;
 }
